hannah: Change device path for second SIM daemon
The RIL daemon (now) seems to use the default device path from the build.prop. Manually pass the argument to libril so that the daemon uses the correct device.

Otherwise Android sees two SIM slots but both correspond to the same physical slot, have the same IMEI and card etc.

Change-Id: I526b3a3d4049ac897b417dcf367f9dedccae5e6d